pyproject.toml
[project]
name = "rag-url"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-genai>=1.21.1",
    "lancedb>=0.24.0",
    "pydantic>=2.11.7",
    "pydantic-ai>=0.3.4",
    "python-dotenv>=1.1.1",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.4",
    "trafilatura>=2.0.0",
]
[tool.setuptools.packages.find]
include = ["rag_url"]
